services/document-extractor/requirements.txt
Flask==3.0.0
google-cloud-documentai==2.20.0
gunicorn==21.2.0
